Add verify that 6 menu items are displayed, carousel buttons for right and left are working, and text of the color is black.
Create three methods: verifyMenuItems(), verifyCarouselButton() and verifyColor(). verifyMenuItems is called in test verifyNumberOfMenuItems()
where it is checked if there are 6 elements. verifyCarouselButton is called in test verifyCarouselButtonForRightAndLeft() where the buttons is checked to see if it works. verifyColor is called int test verifyColorOfTheText() where color of the text is checked, based on the text color defined on the booking website.