Create two test clickImageCheck and selectPriceForPlace
A method has been created clickImage() which scrolls the screen to a specific image, click the image and check if a good page is open. The name of this test is clickImageCheck(). Other test
selectPriceForPlace() contains methods clickImage(), selectAdultsChildrenRooms() where a larger number of people and rooms are placed, for large price, then method clickOnSearchButton(), and setBudget() where a certain minimum and maximum value of the budget is set.