Created a test to select the cheapest accommodation.
A method is created that calls the functions clickImage() which chooses Vrnjacka banja, then sort by low price by calling a function sortByLowPrice(), and choose first property with function clickCheapestProperty(). It is also called scrollUp(), to focus the screen on the top of the page after opening a new tab where it is displayed cheapest property.